+++ admin/partials/tmsm-aquatonic-course-booking-admin-options-page.php
@@ -173,9 +173,7 @@
 			</div>
 		</form>
 
-		<?php
-		$bookings->display();
-		?>
+
 
 		<?php
 	}