+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Store Aquos error in booking entry as _aquos_status 
+* Store Aquos error in booking entry as _aquos_status
+* Bookings list now with phone number and barcode, but displayed on mouseover the full name
 
 ### 1.7.0: July 6th, 2021
 * Refactoring of dashboard calculation, now calculated with cron event