+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.7.1: July 7th, 2021
 * Store Aquos error in booking entry as _aquos_status
 * Bookings list now with phone number and barcode, but displayed on mouseover the full name
 * Bookings list now lists active and noshow bookings by default